Move module from JPMS to Modular
