+++ pom.xml
@@ -48,10 +48,10 @@
         <module>extension</module>
         <module>faces</module>
         <module>http</module>
-        <module>jpms</module>
         <module>maven</module>
         <module>micro</module>
         <module>microprofile</module>
+        <module>modular</module>
         <module>monitor</module>
         <module>naming</module>
         <module>nano</module>
